fix(observability): repair admin Command Center (summary/trends/top/rollups) (#655)
* fix(observability): repair command-center trend rollup refresh

The `telemetry_trend_rollups` materialized view backing the admin Command
Center's 7d/30d trend charts could never be refreshed. Its unique index was
built on an expression (`COALESCE(route, '')`), and PostgreSQL only accepts a
plain-column unique index as the row-matching index for `REFRESH ...
CONCURRENTLY`. So the hourly refresh failed every cycle with "cannot refresh
materialized view concurrently", the view stayed empty since introduction, and
7d/30d charts showed no data. Live (1h/6h/24h) ranges were unaffected — they
read `telemetry_metric_samples` directly.

- Migration recreates the view with `route` as a non-null column
  (`COALESCE(labels->>'http.route', '')`) and a plain-column unique index on
  `(day, metric_name, scope, route)`, which CONCURRENTLY accepts.
- The refresh job now falls back to a blocking `REFRESH` if the concurrent
  refresh can't proceed, so rollups self-heal instead of silently going stale.

Verified against the production schema/data in a rolled-back transaction: the
recreated view populates (222 rows) and the index is plain-column.


* fix(observability): repair command-center summary/trends/top queries

The Command Center showed no data once telemetry accumulated — the endpoints
errored rather than returning empty:

- summary (500) and top-routes/top-errors: their SUM(value_count) over the
  bigint column returns NUMERIC, which failed to decode into i64. Cast the
  aggregates back to ::bigint. (Invisible on an empty DB: SUM over zero rows
  is NULL, which decodes fine — so it only broke in production with data.)
- trends (400): the endpoint used axum's Query (serde_urlencoded), which
  can't parse repeated `metric=` params into a Vec. Switch to axum_extra's
  Query (serde_html_form); enable the axum-extra `query` feature.
- 7d/30d rollups: the materialized view's SUM columns are also cast to
  ::bigint so the rollup trend query decodes.

Adds an integration regression test that seeds samples and asserts the
summary/top-routes/top-errors/trends queries decode with data present — the
coverage gap that let this ship (previous tests ran against an empty DB).

Verified against production data (types now decode as bigint; migration
populates the matview).


---------

+++ server/src/admin/queries.rs
@@ -949,10 +949,12 @@ pub async fn summary_error_and_request_counts(
     from: DateTime<Utc>,
     to: DateTime<Utc>,
 ) -> Result<Option<(Option<i64>, Option<i64>)>, sqlx::Error> {
+    // SUM() over a bigint column returns NUMERIC in PostgreSQL; cast back to
+    // bigint so it decodes into i64 (the sums are small request counts).
     sqlx::query_as::<_, (Option<i64>, Option<i64>)>(
         "SELECT \
-             SUM(CASE WHEN metric_name = 'kaiku_http_errors_total' THEN value_count ELSE 0 END), \
-             SUM(CASE WHEN metric_name = 'kaiku_http_requests_total' THEN value_count ELSE 0 END) \
+             SUM(CASE WHEN metric_name = 'kaiku_http_errors_total' THEN value_count ELSE 0 END)::bigint, \
+             SUM(CASE WHEN metric_name = 'kaiku_http_requests_total' THEN value_count ELSE 0 END)::bigint \
          FROM telemetry_metric_samples \
          WHERE metric_name IN ('kaiku_http_errors_total', 'kaiku_http_requests_total') \
          AND ts >= $1 AND ts <= $2",
